render the datamap container divs in cycle and mount after
this commit moves the containers for the datamaps from the index.html to
our cycle application. we then get the containers after they have been
rendered to the DOM by selecting for them using the observable exposed
for the selector in the DOM driver.

we then get the container in addition to the statistics in the map
driver and either initialize or clear our datamap instance accordingly.

there is an "alernativeView" function that shows how you can toggle the
view on and off and the map driver will do the right thing. (warning:
mounting d3 maps is expensive as hell, so don't actually rip out maps
constantly in a real app just like you wouldn't in a react app)